+++ application/neoform/controller/admin/acl/role/ajax.php
@@ -13,10 +13,25 @@
             switch (http::instance()->slug('action')) {
 
                 case 'insert':
+
+                    if (! auth::instance()->user()->has_resource('admin/acl/role/create')) {
+                        self::show403();
+                        return;
+                    }
+
                     try {
-                        acl\role\api::insert(
+                        $role = acl\role\api::insert(
                             http::instance()->posts()
                         );
+
+                        // Resources
+                        $resource_ids = preg_split('`\s*,\s*`', http::instance()->post('resources'), -1, PREG_SPLIT_NO_EMPTY);
+                        acl\role\resource\api::let(
+                            $role,
+                            new acl\resource\collection($resource_ids)
+                        );
+
+                        $json->role   = $role->export();
                         $json->status = 'good';
                     } catch (input\exception $e) {
                         $json->status  = 'error';
@@ -26,6 +41,12 @@
                     break;
 
                 case 'update':
+
+                    if (! auth::instance()->user()->has_resource('admin/acl/role/edit')) {
+                        self::show403();
+                        return;
+                    }
+
                     try {
                         acl\role\api::update(
                             new acl\role\model(http::instance()->parameter('id')),
@@ -40,6 +61,12 @@
                     break;
 
                 case 'delete':
+
+                    if (! auth::instance()->user()->has_resource('admin/acl/role/delete')) {
+                        self::show403();
+                        return;
+                    }
+
                     try {
                         acl\role\api::delete(
                             new acl\role\model(http::instance()->parameter('id'))
@@ -48,6 +75,27 @@
                     } catch (input\exception $e) {
                         $json->status  = 'error';
                         $json->message = $e->message() ? $e->message() : 'ACL role could not be deleted';
+                        $json->errors  = $e->errors();
+                    }
+                    break;
+
+                case 'update_resources':
+
+                    if (! auth::instance()->user()->has_resource('admin/acl/resource/edit')) {
+                        self::show403();
+                        return;
+                    }
+
+                    try {
+                        $resource_ids = preg_split('`\s*,\s*`', http::instance()->post('resources'), -1, PREG_SPLIT_NO_EMPTY);
+                        acl\role\resource\api::let(
+                            new acl\role\model(http::instance()->parameter('id')),
+                            new acl\resource\collection($resource_ids)
+                        );
+                        $json->status = 'good';
+                    } catch (input\exception $e) {
+                        $json->status  = 'error';
+                        $json->message = $e->message() ? $e->message() : 'Role resources could not be linked';
                         $json->errors  = $e->errors();
                     }
                     break;
