+++ application/neoform/controller/admin/user/ajax.php
@@ -14,11 +14,48 @@
             switch (http::instance()->slug('action')) {
 
                 case 'insert':
+
+                    if (! auth::instance()->user()->has_resource('admin/user/create')) {
+                        self::show403();
+                        return;
+                    }
+
                     try {
-                        user\api::insert(
-                            http::instance()->posts()
-                        );
+                        $http = http::instance();
+
+                        // New user
+                        $user = user\api::admin_insert($http->posts());
+
+                        // Sites
+                        $site_ids = preg_split('`\s*,\s*`', http::instance()->post('sites'), -1, PREG_SPLIT_NO_EMPTY);
+                        if ($site_ids && is_array($site_ids)) {
+                            user\site\api::let(
+                                $user,
+                                new site\collection($site_ids)
+                            );
+                        }
+
+                        // Roles
+                        $role_ids = preg_split('`\s*,\s*`', http::instance()->post('roles'), -1, PREG_SPLIT_NO_EMPTY);
+                        if ($role_ids && is_array($role_ids)) {
+                            user\acl\role\api::let(
+                                $user,
+                                new acl\role\collection($role_ids)
+                            );
+                        }
+
+                        // Groups
+                        $group_ids = preg_split('`\s*,\s*`', http::instance()->post('groups'), -1, PREG_SPLIT_NO_EMPTY);
+                        if ($group_ids && is_array($group_ids)) {
+                            acl\group\user\api::let(
+                                $user,
+                                new acl\group\collection($group_ids)
+                            );
+                        }
+
                         $json->status = 'good';
+                        $json->user   = $user->export(['id', 'email', 'status_id', ]);
+
                     } catch (input\exception $e) {
                         $json->status = 'error';
                         $json->message = $e->message() ?: 'User could not be created';
@@ -27,11 +64,29 @@
                     break;
 
                 case 'update':
+
+                    if (! auth::instance()->user()->has_resource('admin/user/edit')) {
+                        self::show403();
+                        return;
+                    }
+
                     try {
+                        $user = new user\model(http::instance()->parameter('id'));
+
                         user\api::admin_update(
-                            new user\model(http::instance()->parameter('id')),
+                            $user,
                             http::instance()->posts()
                         );
+
+                        // Sites
+                        $site_ids = preg_split('`\s*,\s*`', http::instance()->post('sites'), -1, PREG_SPLIT_NO_EMPTY);
+                        if ($site_ids && is_array($site_ids)) {
+                            user\site\api::let(
+                                $user,
+                                new site\collection($site_ids)
+                            );
+                        }
+
                         $json->status = 'good';
                     } catch (input\exception $e) {
                         $json->status = 'error';
@@ -41,6 +96,12 @@
                     break;
 
                 case 'delete':
+
+                    if (! auth::instance()->user()->has_resource('admin/user/delete')) {
+                        self::show403();
+                        return;
+                    }
+
                     try {
                         user\api::delete(
                             new user\model(http::instance()->parameter('id'))
@@ -54,6 +115,12 @@
                     break;
 
                 case 'update_password':
+
+                    if (! auth::instance()->user()->has_resource('admin/user/edit')) {
+                        self::show403();
+                        return;
+                    }
+
                     try {
                         user\api::admin_password_update(
                             new user\model(http::instance()->parameter('id')),
@@ -68,10 +135,19 @@
                     break;
 
                 case 'update_groups':
+
+                    if (! auth::instance()->user()->has_resource('admin/acl/group/edit')) {
+                        self::show403();
+                        return;
+                    }
+
                     try {
-                        acl\group\user\api::insert(
-                            http::instance()->posts()
+                        $group_ids = preg_split('`\s*,\s*`', http::instance()->post('groups'), -1, PREG_SPLIT_NO_EMPTY);
+                        acl\group\user\api::let(
+                            new user\model(http::instance()->parameter('id')),
+                            new acl\group\collection($group_ids)
                         );
+
                         $json->status = 'good';
                     } catch (input\exception $e) {
                         $json->status = 'error';
@@ -81,10 +157,19 @@
                     break;
 
                 case 'update_roles':
+
+                    if (! auth::instance()->user()->has_resource('admin/acl/role/edit')) {
+                        self::show403();
+                        return;
+                    }
+
                     try {
-                        user\acl\role\api::insert(
-                            http::instance()->posts()
+                        $role_ids = preg_split('`\s*,\s*`', http::instance()->post('roles'), -1, PREG_SPLIT_NO_EMPTY);
+                        user\acl\role\api::let(
+                            new user\model(http::instance()->parameter('id')),
+                            new acl\role\collection($role_ids)
                         );
+
                         $json->status = 'good';
                     } catch (input\exception $e) {
                         $json->status = 'error';
